Moving scripts folder inside package folder (#420)
* Moving scripts folder inside package folder

* Fixing unit tests that rely on config paths

* adding ini files to package data

* Updating docs for scripts move

* Reverting path change for sphinx docs
