+++ machine_common_sense/scripts/cache_addressables.py
