+++ machine_common_sense/scripts/config_level1_debug.ini
