+++ machine_common_sense/scripts/config_level2_debug.ini
