+++ machine_common_sense/scripts/config_no_debug.ini
