+++ machine_common_sense/scripts/config_oracle.ini
