+++ machine_common_sense/scripts/config_with_debug.ini
