+++ machine_common_sense/scripts/run_human_input.py
@@ -1,11 +1,11 @@
 import argparse
 import cmd
 
-from getch_helper import getch
-
 import machine_common_sense as mcs
 from machine_common_sense.goal_metadata import GoalMetadata
 from machine_common_sense.logging_config import LoggingConfig
+
+from .getch_helper import getch
 
 commandList = []
