+++ machine_common_sense/scripts/run_init_scenes.py
