+++ machine_common_sense/scripts/run_last_action.py
@@ -1,4 +1,4 @@
-from runner_script import MultipleFileRunnerScript
+from .runner_script import SingleFileRunnerScript
 
 
 def action_callback(scene_data, step_metadata, runner_script):
@@ -18,7 +18,7 @@ def action_callback(scene_data, step_metadata, runner_script):
 
 
 def main():
-    MultipleFileRunnerScript('Passive Scenes', action_callback)
+    SingleFileRunnerScript('Last Action', action_callback)
 
 
 if __name__ == "__main__":
