+++ machine_common_sense/scripts/run_passive_scenes.py
@@ -1,4 +1,4 @@
-from runner_script import SingleFileRunnerScript
+from .runner_script import MultipleFileRunnerScript
 
 
 def action_callback(scene_data, step_metadata, runner_script):
@@ -18,7 +18,7 @@ def action_callback(scene_data, step_metadata, runner_script):
 
 
 def main():
-    SingleFileRunnerScript('Last Action', action_callback)
+    MultipleFileRunnerScript('Passive Scenes', action_callback)
 
 
 if __name__ == "__main__":
