+++ machine_common_sense/scripts/run_reorientation_scenes.py
