+++ machine_common_sense/scripts/run_scene_timer.py
