+++ machine_common_sense/scripts/run_scene_with_command_file.py
