+++ machine_common_sense/scripts/run_scripts_config_with_history.ini
