+++ machine_common_sense/scripts/sample_command_text.txt
